Functions

Function Reference

Function Reference

AprilTags.jl functions

Wrappers

Index

AprilTags.jl functions

AprilTags.AprilTagDetector — Type.

AprilTagDetector()

Create a default AprilTag detector with tha 36h11 tag family

source

AprilTags.freeDetector! — Function.

freeDetector!(apriltagdetector)

Free the allocated memmory

source

Wrappers

AprilTags.apriltag_detector_create — Function.

apriltag_detector_create()

Create a AprilTag Detector object with all fields set to default value.

source

AprilTags.tag36h11_create — Function.

tag36h11_create()

Create a AprilTag family object for tag36h11 with all fields set to default value.

source

AprilTags.tag36h11_destroy — Function.

tag36h11_destroy(tf)

Destroy the AprilTag family object.

source

AprilTags.apriltag_detector_add_family — Function.

apriltag_detector_add_family(tag_detector, tag_family)

Add a tag family to an AprilTag Detector object. The caller still "owns" the family and a single instance should only be provided to one apriltag detector instance.

source

AprilTags.apriltag_detector_detect — Function.

apriltag_detector_detect(tag_detector, image)

Detect tags from an image and return an array of apriltag_detection_t*. You can use apriltag_detections_destroy to free the array and the detections it contains, or call detection_destroy and zarray_destroy yourself.

source

Index

AprilTags.AprilTagDetector

AprilTags.apriltag_detector_add_family

AprilTags.apriltag_detector_create

AprilTags.apriltag_detector_detect

AprilTags.freeDetector!

AprilTags.tag36h11_create

AprilTags.tag36h11_destroy